requirements.txt
# ── Pentexa Production Dependencies (Vercel) ──
# uvicorn is NOT needed — Vercel provides its own ASGI server.
# celery & redis are kept so import paths don't break at module load time.

fastapi==0.110.0
pydantic==2.6.3
pydantic-settings==2.2.1
sqlalchemy==2.0.28
asyncpg==0.30.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.1
slowapi==0.1.9
redis[hiredis]==5.2.1
bcrypt==4.2.1
celery==5.3.6